feature(order): Prevent deletion of orders with associated products

+++ src/orders/orders.service.ts
@@ -1,13 +1,16 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { ConflictException, forwardRef, Inject, Injectable, NotFoundException } from '@nestjs/common';
 import { CreateOrderDto } from './dto/create-order.dto';
 import { UpdateOrderDto } from './dto/update-order.dto';
 import { InjectModel } from '@nestjs/mongoose';
 import { Order } from './schemas/order.schema';
 import { Model } from 'mongoose';
+import { ProductsService } from 'src/products/products.service';
 
 @Injectable()
 export class OrdersService {
-  constructor(@InjectModel(Order.name) private orderModel: Model<Order>) {}
+  constructor(
+    @InjectModel(Order.name) private orderModel: Model<Order>,
+    @Inject(forwardRef(() => ProductsService)) private productsService: ProductsService) {}
 
   async create(createOrderDto: CreateOrderDto): Promise<Order> {
     const order = new this.orderModel(createOrderDto);
@@ -35,6 +38,16 @@ export class OrdersService {
   }
 
   async remove(id: string): Promise<void> {
+    const order = await this.orderModel.findById(id.toString()).exec();
+    if (!order) {
+      throw new NotFoundException(`Order with ID ${id} not found`);
+    }
+    const products = await this.productsService.findByIds(order.productIds);
+
+    if (products.length > 0) {
+      throw new ConflictException('It is not possible to delete the order because it is linked to one or more products.');
+    }
+    
     const result = await this.orderModel.findByIdAndDelete(id).exec();
     if (!result) {
       throw new NotFoundException(`Order with ID ${id} not found`);
